VisualStudio 10 through 15 are supported
- some minor warnings fixed;
- project files now refer to split platform/config props. Configuration/platform combinations are now stored in platform.props for VS 11.0+ (for VS 10.0 they are conditionally provided from project files);
- platform is auto-detected depending on which VS version used to open the solution.

+++ samples/balls/balls.cpp
@@ -31,7 +31,7 @@ namespace
 		{
 			long long counter;
 			const float dt = 0.3f * (float)stopwatch(_balls_timer);
-			const rect_i area = { 0, 0, surface.width(), surface.height() };
+			const rect_i area = { 0, 0, static_cast<int>(surface.width()), static_cast<int>(surface.height()) };
 
 			_rasterizer.reset();
